Move extension points under strategy for consistency
