+++ src/Transport/Config/ExtensionPoints/AzureServiceBusSanitizationSettings.cs
@@ -1,6 +1,5 @@
 namespace NServiceBus
 {
-    using System;
     using AzureServiceBus;
     using AzureServiceBus.Addressing;
     using Configuration.AdvanceExtensibility;
@@ -65,63 +64,6 @@ namespace NServiceBus
             settings.Set(WellKnownConfigurationKeys.Topology.Addressing.Sanitization.Strategy, typeof(T));
 
             return new AzureServiceBusSanitizationExtensionPoint<T>(settings);
-        }
-
-        public AzureServiceBusSanitizationSettings QueuePathValidation(Func<string, ValidationResult> queuePathValidator)
-        {
-            Guard.AgainstNull(nameof(queuePathValidator), queuePathValidator);
-            settings.Set(WellKnownConfigurationKeys.Topology.Addressing.Sanitization.QueuePathValidator, queuePathValidator);
-            return this;
-        }
-
-        public AzureServiceBusSanitizationSettings TopicPathValidation(Func<string, ValidationResult> topicPathValidator)
-        {
-            Guard.AgainstNull(nameof(topicPathValidator), topicPathValidator);
-            settings.Set(WellKnownConfigurationKeys.Topology.Addressing.Sanitization.TopicPathValidator, topicPathValidator);
-            return this;
-        }
-
-        public AzureServiceBusSanitizationSettings SubscriptionNameValidation(Func<string, ValidationResult> subscriptionNameValidator)
-        {
-            Guard.AgainstNull(nameof(subscriptionNameValidator), subscriptionNameValidator);
-            settings.Set(WellKnownConfigurationKeys.Topology.Addressing.Sanitization.SubscriptionNameValidator, subscriptionNameValidator);
-            return this;
-        }
-
-        public AzureServiceBusSanitizationSettings RuleNameValidation(Func<string, ValidationResult> ruleNameValidator)
-        {
-            Guard.AgainstNull(nameof(ruleNameValidator), ruleNameValidator);
-            settings.Set(WellKnownConfigurationKeys.Topology.Addressing.Sanitization.RuleNameValidator, ruleNameValidator);
-            return this;
-        }
-
-        public AzureServiceBusSanitizationSettings QueuePathSanitization(Func<string, string> queuePathSanitizer)
-        {
-            Guard.AgainstNull(nameof(queuePathSanitizer), queuePathSanitizer);
-            settings.Set(WellKnownConfigurationKeys.Topology.Addressing.Sanitization.QueuePathSanitizer, queuePathSanitizer);
-            return this;
-
-        }
-
-        public AzureServiceBusSanitizationSettings TopicPathSanitization(Func<string, string> topicPathSanitizer)
-        {
-            Guard.AgainstNull(nameof(topicPathSanitizer), topicPathSanitizer);
-            settings.Set(WellKnownConfigurationKeys.Topology.Addressing.Sanitization.TopicPathSanitizer, topicPathSanitizer);
-            return this;
-        }
-
-        public AzureServiceBusSanitizationSettings SubscriptionNameSanitization(Func<string, string> subscriptionNameSanitizer)
-        {
-            Guard.AgainstNull(nameof(subscriptionNameSanitizer), subscriptionNameSanitizer);
-            settings.Set(WellKnownConfigurationKeys.Topology.Addressing.Sanitization.SubscriptionNameSanitizer, subscriptionNameSanitizer);
-            return this;
-        }
-
-        public AzureServiceBusSanitizationSettings RuleNameSanitization(Func<string, string> ruleNameSanitizer)
-        {
-            Guard.AgainstNull(nameof(ruleNameSanitizer), ruleNameSanitizer);
-            settings.Set(WellKnownConfigurationKeys.Topology.Addressing.Sanitization.RuleNameSanitizer, ruleNameSanitizer);
-            return this;
         }
     }
 }
